feat: remove deprecated default generator and evaluation/execution phases
