--- tests/test_evaluation.py
@@ -1,582 +1,0 @@
-# Copyright 2026 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from pathlib import Path
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-from wptgen.config import Config
-from wptgen.models import WorkflowContext
-from wptgen.phases.evaluation import run_test_evaluation
-
-
-@pytest.fixture
-def mock_ui() -> MagicMock:
-  """Fixture that provides a mocked UI provider with a status context manager."""
-  ui = MagicMock()
-  ui.status.return_value.__enter__.return_value = None
-  return ui
-
-
-@pytest.fixture
-def mock_config(tmp_path: Path) -> Config:
-  """Fixture that provides a basic test configuration."""
-  return Config(
-    provider='test',
-    default_model='test-model',
-    api_key='test-key',
-    categories={'lightweight': 'test-model', 'reasoning': 'test-model'},
-    phase_model_mapping={
-      'requirements_extraction': 'reasoning',
-      'coverage_audit': 'reasoning',
-      'generation': 'lightweight',
-      'evaluation': 'lightweight',
-    },
-    wpt_path=str(tmp_path / 'wpt'),
-    cache_path=str(tmp_path / 'cache'),
-    output_dir=str(tmp_path / 'output'),
-  )
-
-
-@pytest.fixture
-def mock_llm() -> MagicMock:
-  """Fixture that provides a mocked LLM client."""
-  llm = MagicMock()
-  llm.generate_content.return_value = 'PASS'
-  llm.model = 'mock-model'
-  return llm
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_pass(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  # Mock style guide and templates
-  style_guide_content = 'Style Guide Content'
-
-  test_path = tmp_path / 'test.html'
-  test_path.write_text('original content')
-  generated_tests = [(test_path, 'original content', '<suggestion></suggestion>')]
-
-  # Targeted patch for the read_text call inside run_test_evaluation
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value=style_guide_content):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  assert test_path.read_text() == 'original content'
-  mock_ui.report_evaluation_result.assert_any_call(test_path.name, success=True)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_correction(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  # Mock style guide and templates
-  style_guide_content = 'Style Guide Content'
-
-  test_path = tmp_path / 'test.html'
-  test_path.write_text('original content')
-  generated_tests = [(test_path, 'original content', '<suggestion></suggestion>')]
-
-  # Mock LLM returning corrected content
-  mock_llm.generate_content.return_value = 'corrected content'
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value=style_guide_content):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  assert test_path.read_text() == 'corrected content\n'
-  mock_ui.report_evaluation_result.assert_any_call(test_path.name, success=True, updated=True)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_reftest_link_fix(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  """Verify that reftest links are corrected during evaluation when files are updated."""
-  context = WorkflowContext(feature_id='reftest-feat')
-  jinja_env = MagicMock()
-
-  test_path = tmp_path / 'reftest-001.html'
-  ref_path = tmp_path / 'reftest-001-ref.html'
-
-  test_path.write_text('<link rel="match" href="old-ref.html">')
-  ref_path.write_text('Reference content')
-
-  suggestion_xml = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_tests = [
-    (test_path, test_path.read_text(), suggestion_xml),
-    (ref_path, ref_path.read_text(), suggestion_xml),
-  ]
-
-  # Mock LLM returning corrected content with a DIFFERENT SUFFIX for the reference
-  # FILE_1: test, FILE_2: ref
-  corrected_content = """
-[FILE_1: .https.html]
-<link rel="match" href="wrong-again.html">
-[/FILE_1]
-[FILE_2: .https.html]
-Updated reference
-[/FILE_2]
-"""
-  mock_llm.generate_content.return_value = corrected_content
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value='Style Guide'):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  # Check that the new files exist and have correct content
-  new_test_path = tmp_path / 'reftest-001.https.html'
-  new_ref_path = tmp_path / 'reftest-001-ref.https.html'
-
-  assert new_test_path.exists()
-  assert new_ref_path.exists()
-
-  # The link in the test should point to the NEW reference filename
-  assert '<link rel="match" href="reftest-001-ref.https.html">' in new_test_path.read_text()
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_with_markdown(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-  style_guide_content = 'Style Guide Content'
-  test_path = tmp_path / 'test_markdown.html'
-  test_path.write_text('original content')
-  generated_tests = [(test_path, 'original content', '<suggestion></suggestion>')]
-
-  # Mock LLM returning content wrapped in markdown blocks
-  mock_llm.generate_content.return_value = '```html\ncorrected markdown content\n```'
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value=style_guide_content):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  assert test_path.read_text() == 'corrected markdown content\n'
-  mock_ui.report_evaluation_result.assert_any_call(test_path.name, success=True, updated=True)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_no_response(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-  test_path = tmp_path / 'test.html'
-  test_path.write_text('original content')
-  generated_tests = [(test_path, 'original content', '<suggestion></suggestion>')]
-
-  # Mock LLM returning empty response (via generate_safe failing)
-  with patch('wptgen.phases.evaluation.generate_safe', return_value=''):
-    with patch('wptgen.phases.evaluation.Path.read_text', return_value='Style Guide'):
-      await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  assert test_path.read_text() == 'original content'
-  mock_ui.report_evaluation_result.assert_called_with(
-    test_path.name, success=False, message=f'No response for evaluation of {test_path.name}.'
-  )
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_reftest_correction(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  # Mock style guide and templates
-  style_guide_content = 'Style Guide Content'
-
-  test_path = tmp_path / 'test.html'
-  ref_path = tmp_path / 'test-ref.html'
-  test_path.write_text('original test content')
-  ref_path.write_text('original ref content')
-
-  suggestion_xml = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_tests = [
-    (test_path, 'original test content', suggestion_xml),
-    (ref_path, 'original ref content', suggestion_xml),
-  ]
-
-  # Mock LLM returning corrected content for BOTH files using suffixes
-  mock_llm.generate_content.return_value = """
-[FILE_1: .html]
-corrected test content
-[/FILE_1]
-
-[FILE_2: .html]
-corrected ref content
-[/FILE_2]
-"""
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value=style_guide_content):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  assert '<link rel="match" href="test-ref.html">' in test_path.read_text()
-  assert 'corrected test content' in test_path.read_text()
-  assert ref_path.read_text() == 'corrected ref content\n'
-  mock_ui.report_evaluation_result.assert_any_call('test.html', success=True, updated=True)
-  mock_ui.report_evaluation_result.assert_any_call('test-ref.html', success=True, updated=True)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_partitioning_logic(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  """Verify evaluation prompt formatting for reftests vs non-reftests."""
-  jinja_env = MagicMock()
-  eval_template_mock = MagicMock()
-  system_template_mock = MagicMock()
-
-  def get_template(name: str) -> MagicMock:
-    if name == 'evaluation.jinja':
-      return eval_template_mock
-    if name == 'evaluation_system.jinja':
-      return system_template_mock
-    return MagicMock()
-
-  jinja_env.get_template.side_effect = get_template
-
-  # 1. Non-Reftest (Single File)
-  js_test_path = tmp_path / 'feat-001.html'
-  js_test_path.write_text('js content')
-
-  generated_tests = [
-    (
-      js_test_path,
-      'js content',
-      '<test_suggestion><test_type>JavaScript Test</test_type></test_suggestion>',
-    )
-  ]
-
-  context = WorkflowContext(feature_id='feat')
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide'):
-    with patch('wptgen.phases.evaluation.generate_safe', return_value='PASS'):
-      await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  # Check eval prompt for non-reftest: should be raw content (no tags)
-  call_args = eval_template_mock.render.call_args.kwargs
-  assert call_args['generated_code_content'] == 'js content'
-
-  eval_template_mock.render.reset_mock()
-
-  # 2. Reftest (Multi File)
-  ref_test_path = tmp_path / 'feat-002.html'
-  ref_test_path.write_text('test content')
-  ref_ref_path = tmp_path / 'feat-002-ref.html'
-  ref_ref_path.write_text('ref content')
-
-  suggestion_xml = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_reftests = [
-    (ref_test_path, 'test content', suggestion_xml),
-    (ref_ref_path, 'ref content', suggestion_xml),
-  ]
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide'):
-    with patch('wptgen.phases.evaluation.generate_safe', return_value='PASS'):
-      await run_test_evaluation(
-        context, mock_config, mock_llm, mock_ui, jinja_env, generated_reftests
-      )
-
-  # Check eval prompt for reftest: should have tags with suffixes
-  call_args = eval_template_mock.render.call_args.kwargs
-  assert '[FILE_1: .html]' in call_args['generated_code_content']
-  assert 'test content' in call_args['generated_code_content']
-  assert '[FILE_2: .html]' in call_args['generated_code_content']
-  assert 'ref content' in call_args['generated_code_content']
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_reftest_rename(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  """Verify that reftest files are renamed if the suffix changes during evaluation."""
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-  style_guide_content = 'Style Guide Content'
-
-  test_path = tmp_path / 'feat-001.html'
-  ref_path = tmp_path / 'feat-001-ref.html'
-  test_path.write_text('old test')
-  ref_path.write_text('old ref')
-
-  suggestion_xml = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_tests = [
-    (test_path, 'old test', suggestion_xml),
-    (ref_path, 'old ref', suggestion_xml),
-  ]
-
-  # Mock LLM returning corrected content with NEW SUFFIXES
-  mock_llm.generate_content.return_value = """
-[FILE_1: .https.html]
-new test content
-[/FILE_1]
-
-[FILE_2: .sub.html]
-new ref content
-[/FILE_2]
-"""
-
-  with patch('wptgen.phases.evaluation.Path.read_text', return_value=style_guide_content):
-    await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  # Verify files were renamed/updated
-  new_test_path = tmp_path / 'feat-001.https.html'
-  new_ref_path = tmp_path / 'feat-001-ref.sub.html'
-
-  assert new_test_path.exists()
-  assert '<link rel="match" href="feat-001-ref.sub.html">' in new_test_path.read_text()
-  assert 'new test content' in new_test_path.read_text()
-  assert not test_path.exists()
-
-  assert new_ref_path.exists()
-  assert new_ref_path.read_text() == 'new ref content\n'
-  assert not ref_path.exists()
-
-  mock_ui.report_evaluation_result.assert_any_call(
-    'feat-001.https.html', success=True, updated=True
-  )
-  mock_ui.report_evaluation_result.assert_any_call(
-    'feat-001-ref.sub.html', success=True, updated=True
-  )
-
-
-@pytest.mark.asyncio
-async def test_run_wpt_lint_error(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  from wptgen.phases.evaluation import _run_wpt_lint
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'Mock lint error', b'')
-  mock_process.returncode = 1
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  result = await _run_wpt_lint(Path('/mock/wpt/test.html'), Path('/mock/wpt'))
-  assert result == 'Mock lint error'
-
-
-@pytest.mark.asyncio
-async def test_run_wpt_lint_exception(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  from wptgen.phases.evaluation import _run_wpt_lint
-
-  mocker.patch('asyncio.create_subprocess_exec', side_effect=Exception('Mock exec error'))
-  result = await _run_wpt_lint(Path('/mock/wpt/test.html'), Path('/mock/wpt'))
-  assert result is not None
-  assert result is not None
-
-  assert 'Failed to run ./wpt lint: Mock exec error' in result
-
-
-@pytest.mark.asyncio
-async def test_run_wpt_lint_relative_wpt_dir(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  import os
-
-  from wptgen.phases.evaluation import _run_wpt_lint
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'', b'')
-  mock_process.returncode = 0
-  mock_exec = mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  wpt_dir = tmp_path / 'wpt'
-  wpt_dir.mkdir(parents=True, exist_ok=True)
-  test_file = wpt_dir / 'test.html'
-  test_file.touch()
-
-  original_cwd = Path.cwd()
-  try:
-    os.chdir(tmp_path)
-    # The relative path from tmp_path to tmp_path/wpt is simply 'wpt'
-    rel_wpt_dir = Path('wpt')
-
-    # Run lint with absolute test_file path and relative wpt_dir path
-    result = await _run_wpt_lint(test_file.absolute(), rel_wpt_dir)
-
-    assert result is None
-    mock_exec.assert_called_once()
-    # Assert that the relative path passed to wpt lint is correct
-    assert mock_exec.call_args[0][2] == 'test.html'
-  finally:
-    os.chdir(original_cwd)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_non_reftest_multi_file(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  test1 = Path(mock_config.wpt_path) / 'test1.html'
-  test1.parent.mkdir(parents=True, exist_ok=True)
-  test2 = tmp_path / 'test2.html'
-  test1.write_text('content1')
-  test2.write_text('content2')
-
-  suggestion = '<test_suggestion><test_type>JavaScript Test</test_type></test_suggestion>'
-  generated_tests = [(test1, 'content1', suggestion), (test2, 'content2', suggestion)]
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'Lint error', b'')
-  mock_process.returncode = 1
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  mocker.patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide')
-  mocker.patch('wptgen.phases.evaluation.generate_safe', return_value='PASS')
-
-  await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  mock_ui.report_evaluation_result.assert_any_call('test1.html, test2.html', success=True)
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_reftest_single_correction(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  test = tmp_path / 'reftest.html'
-  ref = tmp_path / 'reftest-ref.html'
-  test.write_text('test content')
-  ref.write_text('ref content')
-
-  suggestion = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_tests = [(test, 'test content', suggestion), (ref, 'ref content', suggestion)]
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'Lint error', b'')
-  mock_process.returncode = 1
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  # Mock LLM returning a single string WITHOUT [FILE_1] tags
-  mocker.patch('wptgen.phases.evaluation.generate_safe', return_value='Just one file output')
-  mocker.patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide')
-
-  await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  mock_ui.report_evaluation_result.assert_any_call(
-    'reftest.html, reftest-ref.html',
-    success=False,
-    message='Received single-file correction for multi-file test reftest.html, reftest-ref.html. Skipping.',
-  )
-
-
-@pytest.mark.asyncio
-async def test_run_wpt_lint_success_and_no_errors(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  test1 = Path(mock_config.wpt_path) / 'test1.html'
-  test1.parent.mkdir(parents=True, exist_ok=True)
-  test1.write_text('content1')
-
-  suggestion = '<test_suggestion><test_type>JavaScript Test</test_type></test_suggestion>'
-  generated_tests = [(test1, 'content1', suggestion)]
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'', b'')
-  mock_process.returncode = 0
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  mocker.patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide')
-  mocker.patch('wptgen.phases.evaluation.generate_safe', return_value='PASS')
-
-  await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  mock_ui.print.assert_any_call('[green]No lint errors found for test1.html![/green]')
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_reftest_three_files(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  test1 = Path(mock_config.wpt_path) / 'test1.html'
-  test1.parent.mkdir(parents=True, exist_ok=True)
-  test2 = Path(mock_config.wpt_path) / 'test2.html'
-  test3 = Path(mock_config.wpt_path) / 'test3.html'
-  test1.write_text('content1')
-  test2.write_text('content2')
-  test3.write_text('content3')
-
-  suggestion = '<test_suggestion><test_type>Reftest</test_type></test_suggestion>'
-  generated_tests = [
-    (test1, 'content1', suggestion),
-    (test2, 'content2', suggestion),
-    (test3, 'content3', suggestion),
-  ]
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'Lint error', b'')
-  mock_process.returncode = 1
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  mocker.patch('wptgen.phases.evaluation.Path.read_text', return_value='Guide')
-  mocker.patch('wptgen.phases.evaluation.generate_safe', return_value='PASS')
-
-  await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  mock_ui.report_evaluation_result.assert_any_call(
-    'test1.html, test2.html, test3.html', success=True
-  )
-
-
-@pytest.mark.asyncio
-async def test_run_test_evaluation_single_file_with_multifile_block(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  context = WorkflowContext(feature_id='feat')
-  jinja_env = MagicMock()
-
-  test_path = tmp_path / 'test.html'
-  test_path.write_text('original')
-
-  suggestion = '<test_suggestion><test_type>JavaScript Test</test_type></test_suggestion>'
-  generated_tests = [(test_path, 'original', suggestion)]
-
-  mock_process = mocker.AsyncMock()
-  mock_process.communicate.return_value = (b'Lint error', b'')
-  mock_process.returncode = 1
-  mocker.patch('asyncio.create_subprocess_exec', return_value=mock_process)
-
-  # LLM responds with a multi-file block for a single file test, and changes the suffix!
-  mock_response = '[FILE_1: .https.html]\nupdated content\n[/FILE_1]'
-  mocker.patch('wptgen.phases.evaluation.generate_safe', return_value=mock_response)
-
-  await run_test_evaluation(context, mock_config, mock_llm, mock_ui, jinja_env, generated_tests)
-
-  new_test_path = tmp_path / 'test.https.html'
-  assert new_test_path.exists()
-  assert new_test_path.read_text() == 'updated content\n'
-  assert not test_path.exists()
-
-  mock_ui.report_evaluation_result.assert_called_with(
-    new_test_path.name, success=True, updated=True
-  )
