+++ tests/test_main.py
@@ -41,7 +41,6 @@ def mock_config(tmp_path: Path) -> Config:
       'requirements_extraction': 'reasoning',
       'coverage_audit': 'reasoning',
       'generation': 'lightweight',
-      'evaluation': 'lightweight',
     },
     wpt_path=str(tmp_path / 'wpt'),
     cache_path=str(tmp_path / 'cache'),
@@ -106,9 +105,6 @@ def test_generate_success(mocker: MockerFixture, mock_config: Config) -> None:
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -116,7 +112,6 @@ def test_generate_success(mocker: MockerFixture, mock_config: Config) -> None:
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
   mock_engine_class.assert_called_once()
   # Verify config was passed correctly
@@ -157,9 +152,6 @@ def test_generate_show_responses(mocker: MockerFixture, mock_config: Config) -> 
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -167,7 +159,6 @@ def test_generate_show_responses(mocker: MockerFixture, mock_config: Config) -> 
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -204,9 +195,6 @@ def test_generate_yes_tokens(mocker: MockerFixture, mock_config: Config) -> None
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -214,7 +202,6 @@ def test_generate_yes_tokens(mocker: MockerFixture, mock_config: Config) -> None
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -251,9 +238,6 @@ def test_generate_suggestions_only(mocker: MockerFixture, mock_config: Config) -
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -261,7 +245,6 @@ def test_generate_suggestions_only(mocker: MockerFixture, mock_config: Config) -
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -298,9 +281,6 @@ def test_generate_max_retries(mocker: MockerFixture, mock_config: Config) -> Non
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -308,7 +288,6 @@ def test_generate_max_retries(mocker: MockerFixture, mock_config: Config) -> Non
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -345,9 +324,6 @@ def test_generate_detailed_requirements(mocker: MockerFixture, mock_config: Conf
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -355,94 +331,6 @@ def test_generate_detailed_requirements(mocker: MockerFixture, mock_config: Conf
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
-  )
-
-
-def test_generate_skip_evaluation(mocker: MockerFixture, mock_config: Config) -> None:
-  """Test that the --skip-evaluation/--no-eval flag is correctly passed to load_config."""
-  mock_load_config = mocker.patch('wptgen.main.load_config', return_value=mock_config)
-  mocker.patch('wptgen.main.WPTGenEngine')
-
-  # Run with --skip-evaluation
-  result = runner.invoke(app, ['generate', 'grid', '--skip-evaluation'])
-  assert result.exit_code == 0
-  mock_load_config.assert_called_with(
-    config_path=DEFAULT_CONFIG_PATH,
-    provider_override=None,
-    wpt_dir_override=None,
-    output_dir_override=None,
-    show_responses=False,
-    yes_tokens_override=False,
-    yes_tests_override=False,
-    yes_cache_override=False,
-    no_cache_override=False,
-    suggestions_only=False,
-    brief_suggestions=False,
-    resume_override=False,
-    resume_from_override=None,
-    state_dir_override=None,
-    max_retries_override=3,
-    timeout_override=600,
-    spec_urls_override=None,
-    feature_description_override=None,
-    detailed_requirements_override=False,
-    draft_override=False,
-    single_prompt_requirements_override=False,
-    use_lightweight_override=False,
-    use_reasoning_override=False,
-    skip_evaluation_override=True,
-    skip_execution_override=False,
-    generator='default',
-    agentic_generation_override=False,
-    agentic_yolo_override=False,
-    tentative_override=False,
-    save_traces_override=False,
-    max_parallel_requests_override=None,
-    temperature_override=None,
-    include_thoughts_override=False,
-    wpt_binary_override=None,
-  )
-
-  # Run with --no-eval alias
-  mock_load_config.reset_mock()
-  result = runner.invoke(app, ['generate', 'grid', '--no-eval'])
-  assert result.exit_code == 0
-  mock_load_config.assert_called_with(
-    config_path=DEFAULT_CONFIG_PATH,
-    provider_override=None,
-    wpt_dir_override=None,
-    output_dir_override=None,
-    show_responses=False,
-    yes_tokens_override=False,
-    yes_tests_override=False,
-    yes_cache_override=False,
-    no_cache_override=False,
-    suggestions_only=False,
-    brief_suggestions=False,
-    resume_override=False,
-    resume_from_override=None,
-    state_dir_override=None,
-    max_retries_override=3,
-    timeout_override=600,
-    spec_urls_override=None,
-    feature_description_override=None,
-    detailed_requirements_override=False,
-    draft_override=False,
-    single_prompt_requirements_override=False,
-    use_lightweight_override=False,
-    use_reasoning_override=False,
-    skip_evaluation_override=True,
-    skip_execution_override=False,
-    generator='default',
-    agentic_generation_override=False,
-    agentic_yolo_override=False,
-    tentative_override=False,
-    save_traces_override=False,
-    max_parallel_requests_override=None,
-    temperature_override=None,
-    include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -510,9 +398,6 @@ def test_generate_spec_urls(mocker: MockerFixture, mock_config: Config) -> None:
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -520,7 +405,6 @@ def test_generate_spec_urls(mocker: MockerFixture, mock_config: Config) -> None:
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -557,9 +441,6 @@ def test_generate_description(mocker: MockerFixture, mock_config: Config) -> Non
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -567,7 +448,6 @@ def test_generate_description(mocker: MockerFixture, mock_config: Config) -> Non
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -604,9 +484,6 @@ def test_generate_resume(mocker: MockerFixture, mock_config: Config) -> None:
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -614,7 +491,6 @@ def test_generate_resume(mocker: MockerFixture, mock_config: Config) -> None:
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -651,9 +527,6 @@ def test_generate_use_lightweight(mocker: MockerFixture, mock_config: Config) ->
     single_prompt_requirements_override=False,
     use_lightweight_override=True,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -661,7 +534,6 @@ def test_generate_use_lightweight(mocker: MockerFixture, mock_config: Config) ->
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -698,9 +570,6 @@ def test_generate_use_reasoning(mocker: MockerFixture, mock_config: Config) -> N
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=True,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -708,7 +577,6 @@ def test_generate_use_reasoning(mocker: MockerFixture, mock_config: Config) -> N
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -745,9 +613,6 @@ def test_generate_single_prompt_requirements(mocker: MockerFixture, mock_config:
     single_prompt_requirements_override=True,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -755,7 +620,6 @@ def test_generate_single_prompt_requirements(mocker: MockerFixture, mock_config:
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -792,9 +656,6 @@ def test_generate_max_parallel_requests(mocker: MockerFixture, mock_config: Conf
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -802,7 +663,6 @@ def test_generate_max_parallel_requests(mocker: MockerFixture, mock_config: Conf
     max_parallel_requests_override=5,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -1072,8 +932,6 @@ def test_audit_success(mocker: MockerFixture, mock_config: Config) -> None:
   mock_load_config.assert_called_once()
   kwargs = mock_load_config.call_args.kwargs
   assert kwargs['suggestions_only'] is True
-  assert kwargs['skip_evaluation_override'] is True
-  assert kwargs['skip_execution_override'] is True
   assert kwargs['provider_override'] == 'gemini'
 
   mock_engine_instance.run_workflow.assert_called_once_with('grid')
@@ -1111,9 +969,6 @@ def test_generate_draft(mocker: MockerFixture, mock_config: Config) -> None:
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -1121,7 +976,6 @@ def test_generate_draft(mocker: MockerFixture, mock_config: Config) -> None:
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -1209,93 +1063,6 @@ def test_config_set_command_types(mocker: MockerFixture) -> None:
     assert data['temperature'] == 0.5
 
 
-def test_generate_skip_execution(mocker: MockerFixture, mock_config: Config) -> None:
-  """Test that the --skip-execution/--no-exec flag is correctly passed to load_config."""
-  mock_load_config = mocker.patch('wptgen.main.load_config', return_value=mock_config)
-  mocker.patch('wptgen.main.WPTGenEngine')
-
-  # Run with --skip-execution
-  result = runner.invoke(app, ['generate', 'grid', '--skip-execution'])
-  assert result.exit_code == 0
-  mock_load_config.assert_called_with(
-    config_path=DEFAULT_CONFIG_PATH,
-    provider_override=None,
-    wpt_dir_override=None,
-    output_dir_override=None,
-    show_responses=False,
-    yes_tokens_override=False,
-    yes_tests_override=False,
-    yes_cache_override=False,
-    no_cache_override=False,
-    suggestions_only=False,
-    brief_suggestions=False,
-    resume_override=False,
-    resume_from_override=None,
-    state_dir_override=None,
-    max_retries_override=3,
-    timeout_override=600,
-    spec_urls_override=None,
-    feature_description_override=None,
-    detailed_requirements_override=False,
-    draft_override=False,
-    single_prompt_requirements_override=False,
-    use_lightweight_override=False,
-    use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=True,
-    generator='default',
-    agentic_generation_override=False,
-    agentic_yolo_override=False,
-    tentative_override=False,
-    save_traces_override=False,
-    max_parallel_requests_override=None,
-    temperature_override=None,
-    include_thoughts_override=False,
-    wpt_binary_override=None,
-  )
-
-  # Run with --no-exec alias
-  mock_load_config.reset_mock()
-  result = runner.invoke(app, ['generate', 'grid', '--no-exec'])
-  assert result.exit_code == 0
-  mock_load_config.assert_called_with(
-    config_path=DEFAULT_CONFIG_PATH,
-    provider_override=None,
-    wpt_dir_override=None,
-    output_dir_override=None,
-    show_responses=False,
-    yes_tokens_override=False,
-    yes_tests_override=False,
-    yes_cache_override=False,
-    no_cache_override=False,
-    suggestions_only=False,
-    brief_suggestions=False,
-    resume_override=False,
-    resume_from_override=None,
-    state_dir_override=None,
-    max_retries_override=3,
-    timeout_override=600,
-    spec_urls_override=None,
-    feature_description_override=None,
-    detailed_requirements_override=False,
-    draft_override=False,
-    single_prompt_requirements_override=False,
-    use_lightweight_override=False,
-    use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=True,
-    generator='default',
-    agentic_generation_override=False,
-    agentic_yolo_override=False,
-    tentative_override=False,
-    save_traces_override=False,
-    max_parallel_requests_override=None,
-    temperature_override=None,
-    include_thoughts_override=False,
-    wpt_binary_override=None,
-  )
-
-
 def test_generate_agentic_generation(mocker: MockerFixture, mock_config: Config) -> None:
   """Test that the --agentic-generation flag is correctly passed to load_config."""
   mock_load_config = mocker.patch('wptgen.main.load_config', return_value=mock_config)
@@ -1328,9 +1095,6 @@ def test_generate_agentic_generation(mocker: MockerFixture, mock_config: Config)
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=True,
     agentic_yolo_override=False,
     tentative_override=False,
@@ -1338,7 +1102,6 @@ def test_generate_agentic_generation(mocker: MockerFixture, mock_config: Config)
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
 
 
@@ -1374,9 +1137,6 @@ def test_generate_agentic_yolo(mocker: MockerFixture, mock_config: Config) -> No
     single_prompt_requirements_override=False,
     use_lightweight_override=False,
     use_reasoning_override=False,
-    skip_evaluation_override=False,
-    skip_execution_override=False,
-    generator='default',
     agentic_generation_override=False,
     agentic_yolo_override=True,
     tentative_override=False,
@@ -1384,7 +1144,6 @@ def test_generate_agentic_yolo(mocker: MockerFixture, mock_config: Config) -> No
     max_parallel_requests_override=None,
     temperature_override=None,
     include_thoughts_override=False,
-    wpt_binary_override=None,
   )
