+++ tests/test_phase_utils.py
@@ -54,7 +54,6 @@ def mock_config(tmp_path: Path) -> Config:
       'requirements_extraction': 'reasoning',
       'coverage_audit': 'reasoning',
       'generation': 'lightweight',
-      'evaluation': 'lightweight',
     },
     wpt_path=str(tmp_path / 'wpt'),
     cache_path=str(tmp_path / 'cache'),
