+++ tests/test_phases.py
@@ -57,7 +57,6 @@ def mock_config(tmp_path: Path) -> Config:
       'requirements_extraction': 'reasoning',
       'coverage_audit': 'reasoning',
       'generation': 'lightweight',
-      'evaluation': 'lightweight',
     },
     wpt_path=str(tmp_path / 'wpt'),
     cache_path=str(tmp_path / 'cache'),
@@ -136,6 +135,7 @@ async def test_run_context_assembly_unregistered_with_params(
 
   with (
     patch('wptgen.phases.context_assembly.fetch_feature_yaml', return_value=None),
+    patch('wptgen.phases.context_assembly.fetch_mdn_urls', return_value=[]),
     patch('wptgen.phases.context_assembly.fetch_and_extract_text', return_value='Spec Content'),
     patch('wptgen.phases.context_assembly.find_feature_tests', return_value=[]),
     patch('wptgen.phases.context_assembly.gather_local_test_context', return_value=WPTContext()),
@@ -444,41 +444,6 @@ async def test_run_test_generation_no_suggestions(
 
 
 @pytest.mark.asyncio
-async def test_run_test_generation_reftest_link_fix(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  """Verify that reftest links are corrected during generation."""
-  mock_config.output_dir = str(tmp_path)
-  context = WorkflowContext(
-    feature_id='reftest-feat',
-    audit_response='<status>TESTS_NEEDED</status><test_suggestion><test_type>Reftest</test_type><name>my-test</name></test_suggestion>',
-    metadata=FeatureMetadata('Feat', 'Desc', ['http://spec']),
-  )
-  jinja_env = MagicMock()
-  jinja_env.get_template.return_value.render.return_value = 'Mock Template'
-
-  reftest_content = """
-[FILE_1: .html]
-<link rel="match" href="wrong-ref.html">
-[/FILE_1]
-[FILE_2: .html]
-Reference content
-[/FILE_2]
-"""
-
-  with patch('wptgen.phases.generation.generate_safe', return_value=reftest_content):
-    res = await run_test_generation(context, mock_config, mock_llm, mock_ui, jinja_env)
-
-  assert len(res) == 2
-  test_path, test_content, _ = res[0]
-  ref_path, ref_content, _ = res[1]
-
-  assert test_path.name == 'reftest-feat-001.html'
-  assert ref_path.name == 'reftest-feat-001-ref.html'
-  assert '<link rel="match" href="reftest-feat-001-ref.html">' in test_content
-
-
-@pytest.mark.asyncio
 async def test_run_test_generation_none_selected(
   mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock
 ) -> None:
@@ -499,32 +464,6 @@ async def test_run_test_generation_none_selected(
 
   assert res == []
   mock_ui.warning.assert_any_call('No tests selected. Exiting.')
-
-
-@pytest.mark.asyncio
-async def test_run_test_generation_failure(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock, tmp_path: Path
-) -> None:
-  """Test test generation failure when LLM call fails."""
-  suggestion_xml = (
-    '<test_suggestion><title>T1</title><description>D1</description></test_suggestion>'
-  )
-  context = WorkflowContext(
-    feature_id='feat',
-    metadata=FeatureMetadata('Feat', 'Desc', ['http://spec']),
-    audit_response=suggestion_xml,
-  )
-  jinja_env = MagicMock()
-  jinja_env.get_template.return_value.render.return_value = 'Prompt'
-
-  mock_ui.confirm.return_value = True
-  # Mock generation_safe to return empty string (failure)
-  with patch('wptgen.phases.generation.generate_safe', return_value=''):
-    with patch('wptgen.phases.generation.Path.read_text', return_value='Style Guide'):
-      res = await run_test_generation(context, mock_config, mock_llm, mock_ui, jinja_env)
-
-  assert res == []
-  mock_ui.report_generation_summary.assert_called_once_with([])
 
 
 @pytest.mark.asyncio
@@ -594,62 +533,6 @@ async def test_run_test_generation_agentic(
   )
 
 
-@pytest.mark.asyncio
-async def test_run_test_generation_displays_worksheet(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock
-) -> None:
-  """Verify that the audit worksheet is displayed before suggestions."""
-  suggestion_xml = (
-    '<test_suggestion><title>T1</title><description>D1</description></test_suggestion>'
-  )
-  audit_response = f'<audit_worksheet>R1: Req 1 -> [COVERED by test.html]\nR2: Req 2 -> [UNCOVERED]</audit_worksheet>\n{suggestion_xml}'
-  context = WorkflowContext(
-    feature_id='feat',
-    metadata=FeatureMetadata('Feat', 'Desc', ['http://spec']),
-    audit_response=audit_response,
-  )
-  jinja_env = MagicMock()
-  jinja_env.get_template.return_value.render.return_value = 'Generated Content'
-
-  mock_ui.confirm.return_value = False  # Stop after displaying worksheet and suggestions
-
-  with (
-    patch('wptgen.phases.generation.confirm_prompts', return_value=None),
-    patch('wptgen.phases.generation.Path.read_text', return_value='Style Guide Content'),
-  ):
-    await run_test_generation(context, mock_config, mock_llm, mock_ui, jinja_env)
-
-  # Check that the worksheet was displayed
-  mock_ui.report_audit_worksheet.assert_called_once()
-
-
-@pytest.mark.asyncio
-async def test_run_test_generation_yes_tests(
-  mock_config: Config, mock_ui: MagicMock, mock_llm: MagicMock
-) -> None:
-  """Verify that yes_tests config bypasses user confirmation."""
-  suggestion_xml = (
-    '<test_suggestion><title>T1</title><description>D1</description></test_suggestion>'
-  )
-  context = WorkflowContext(
-    feature_id='feat',
-    metadata=FeatureMetadata('Feat', 'Desc', ['http://spec']),
-    audit_response=suggestion_xml,
-  )
-  jinja_env = MagicMock()
-  mock_config.yes_tests = True
-
-  with (
-    patch('wptgen.phases.generation.confirm_prompts', return_value=None),
-    patch('wptgen.phases.generation.Path.read_text', return_value='Style Guide Content'),
-    patch('wptgen.phases.generation._generate_and_save', return_value=[]),
-  ):
-    await run_test_generation(context, mock_config, mock_llm, mock_ui, jinja_env)
-
-  # Check that the confirmation prompt was NOT called
-  mock_ui.confirm.assert_not_called()
-
-
 def test_partition_requirements_xml() -> None:
   # Empty or no tags
   assert partition_requirements_xml('') == []
@@ -711,39 +594,6 @@ R3: COVERED
   assert 'R3: COVERED' in combined
   assert '<test_suggestion>Suggestion 1</test_suggestion>' in combined
   assert '<test_suggestion>Suggestion 2</test_suggestion>' in combined
-
-
-@pytest.mark.asyncio
-async def test_process_generation_result_single_file_fallback(
-  mocker: MagicMock, mock_config: Config, mock_ui: MagicMock, tmp_path: Path
-) -> None:
-  from wptgen.phases.generation import _generate_and_save
-
-  mock_config.output_dir = str(tmp_path)
-
-  content = '```html\nfallback html content\n```'
-  suggestion_xml = '<test_suggestion></test_suggestion>'
-
-  mocker.patch('wptgen.phases.generation.parse_multi_file_response', return_value=[])
-  mock_llm = MagicMock()
-  mocker.patch('wptgen.phases.generation.generate_safe', return_value=content)
-
-  results = await _generate_and_save(
-    prompt='mock prompt',
-    root_name='test-root',
-    suggestion_xml=suggestion_xml,
-    llm=mock_llm,
-    ui=mock_ui,
-    config=mock_config,
-  )
-
-  assert len(results) == 1
-  assert results[0][0] == tmp_path / 'test-root.html'
-  assert 'fallback html content' in results[0][1]
-
-  mock_ui.report_test_generated.assert_called_with(
-    'test-root', success=True, path=results[0][0], fallback=True
-  )
 
 
 @pytest.mark.asyncio
@@ -820,7 +670,6 @@ async def test_run_test_generation_adk(
   from wptgen.models import WorkflowContext
   from wptgen.phases.generation import run_test_generation
 
-  mock_config.generator = 'adk'
   suggestion_xml = (
     '<test_suggestion><title>T1</title><description>D1</description></test_suggestion>'
   )
@@ -850,7 +699,6 @@ async def test_generate_adk_loop(mock_config: Config, mock_ui: MagicMock) -> Non
   from wptgen.models import WorkflowContext
   from wptgen.phases.generation import _generate_adk_loop
 
-  mock_config.generator = 'adk'
   mock_config.output_dir = 'test_dir'
   mock_config.brief_suggestions = False
