+++ tests/test_resume.py
@@ -39,7 +39,6 @@ def mock_config(tmp_path: Path) -> Config:
       'requirements_extraction': 'reasoning',
       'coverage_audit': 'reasoning',
       'generation': 'lightweight',
-      'evaluation': 'lightweight',
     },
     wpt_path=str(tmp_path / 'wpt'),
     cache_path=str(tmp_path / '.wpt-gen-cache'),
