+++ wpt-gen.yml
@@ -1,8 +1,6 @@
 version: "1"
 default_provider: gemini
 wpt_path: ../wpt
-wpt_browser: chrome
-wpt_channel: canary
 providers:
   gemini:
     default_model: gemini-3.1-pro-preview
@@ -24,4 +22,3 @@ phase_model_mapping:
   requirements_extraction: reasoning
   coverage_audit: reasoning
   generation: lightweight
-  evaluation: lightweight
