+++ wptgen/phases/coverage_audit.py
@@ -115,9 +115,7 @@ async def run_coverage_audit(
   spec_urls = context.metadata.specs if context.metadata and context.metadata.specs else []
 
   audit_system_prompt = jinja_env.get_template('coverage_audit_system.jinja').render(
-    brief_suggestions=(
-      config.brief_suggestions or config.agentic_generation or config.generator == 'adk'
-    ),
+    brief_suggestions=True,
     spec_urls=spec_urls,
   )
