--- wptgen/templates/correction.jinja
@@ -1,9 +1,0 @@
-Fix the following failing test.
-
-<error_log>
-{{ error_log }}
-</error_log>
-
-<test_source_code>
-{{ test_source_code }}
-</test_source_code>
