--- wptgen/templates/evaluation.jinja
@@ -1,15 +1,0 @@
-Review the following generated WPT code against its source blueprint.
-
-<blueprint>
-{{test_suggestion_xml}}
-</blueprint>
-
-<generated_code>
-{{generated_code_content}}
-</generated_code>
-{% if lint_errors %}
-<lint_errors>
-The following errors were reported by `./wpt lint`:
-{{lint_errors}}
-</lint_errors>
-{% endif %}
