--- wptgen/templates/test_generation.jinja
@@ -1,13 +1,0 @@
-Generate a single WPT file for the feature "{{feature_name}}" ({{feature_description}}) based on the following blueprint:
-
-{{ test_suggestion_xml_block }}
-
-{% if specs -%}
-# Specification Context
-The exact API definitions, interfaces, and expected behaviors for this feature:
-{% for url, content in specs.items() -%}
-<spec_document url="{{url}}">
-{{ content }}
-</spec_document>
-{%- endfor %}
-{%- endif %}
